Moved patch code to pkg/patch
